Move AppearanceAnim into subcomponents folder
